Moved Components to Their Own Files
This promotes readability for those looking at our projects
