+++ StudentWebsite/src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Container, Row } from "reactstrap";
-import CurrentSound from "./components/CurrentSound";
-import Analytics from "./components/Analytics";
+import CurrentSound from "./components/CurrentSound/CurrentSound";
+import Analytics from "./components/Analytics/Analytics";
 import "bootstrap/dist/css/bootstrap.min.css";
